Backend/requirements.txt
beautifulsoup4==4.9.3
Django==3.2.5
django-cors-headers==3.7.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
nltk==3.6.2
psycopg2==2.9.1
